Add helpers module and import tasks
