+++ flows/atd-service-bot/issues_to_socrata.py
@@ -15,43 +15,11 @@ to an Open Data Portal dataset (AKA Socrata)
 """
 
 import docker
-import json
-
 
 # Prefect
 from prefect import flow, task, get_run_logger
-from prefect.blocks.system import Secret
 
-
-# Task to pull the latest Docker image
-@task(
-    name="pull_docker_image",
-    retries=1,
-    retry_delay_seconds=5,
-)
-def pull_docker_image(docker_tag):
-    logger = get_run_logger()
-
-    docker_image = f"atddocker/atd-service-bot:{docker_tag}"
-    client = docker.from_env()
-    client.images.pull("atddocker/atd-service-bot", tag=docker_tag)
-    logger.info(f"Docker Images Pulled, using: {docker_image}")
-    return docker_image
-
-
-# Get the environment variables based on the given environment
-@task(name="get_env_vars", timeout_seconds=60)
-def get_env_vars(env):
-    logger = get_run_logger()
-
-    # Environment Variables stored in secret block in Prefect
-    secret_block = Secret.load(f"atd-service-bot-{env}")
-    env_vars_json_string = secret_block.get()
-    environment_variables = json.loads(env_vars_json_string)
-
-    logger.info(f"Received Prefect Environment Variables for: {env}")
-    return environment_variables
-
+from helpers import get_env_vars, pull_docker_image
 
 # Issues to Socrata
 @task(name="sending_issues_to_socrata", timeout_seconds=60)
